+++ web/src/pages/Main/Main.tsx
@@ -1,19 +1,25 @@
 import React from 'react'
 import { withRouter } from "react-router";
+import {useFetchAll} from '../../API/fetchRequests';
 import UserComponent from '../../components/UserComponent/UserComponent'
 import ErrorMessageComponent from '../../components/ErrorMessageComponent/ErrorMessageComponent';
-import {useFetchAll} from '../../API/fetchRequests';
+import LoadingComponent from '../../components/LoadingComponent/LoadingComponent';
+
 
 function Main() {
-	const [data, error] = useFetchAll([]) 
+	const [userDataAll, error, loading] = useFetchAll() 
 
 	if(!!error){
 		return <ErrorMessageComponent message = {error}/>
 	}
 
+	if (loading) {
+		return <LoadingComponent/>
+	}
+
 	return (
 		<div>
-			<UserComponent  userDataAll = {data}/>
+			<UserComponent  userDataAll = {userDataAll}/>
 		</div>
 	)
 }
