+++ web/src/pages/User/User.tsx
@@ -1,30 +1,27 @@
 import React from 'react'
 import { withRouter, useParams } from "react-router";
+import {useFetchUser} from '../../API/fetchRequests';
 import UserDetails from '../../components/UserDetails/UserDetails';
 import ErrorMessageComponent from '../../components/ErrorMessageComponent/ErrorMessageComponent';
-import {useFetchUser} from '../../API/fetchRequests';
+import LoadingComponent from '../../components/LoadingComponent/LoadingComponent';
 
 interface Params {
 	id: string;
 }
 
-const defaultUser ={
-	"id": "",
-	"first_name": "",
-	"last_name": "",
-	"email": "",
-	"favorite_color": "",
-	"number_of_pets": 0
-}
-
 function User() {
 	const params:Params = useParams()
-	const [data, error] = useFetchUser(defaultUser,params.id)
+	const [userData, error, loading] = useFetchUser(params.id)
+	
 	if(!!error) return <ErrorMessageComponent message = {error}/>
+
+	if (loading) {
+		return <LoadingComponent/>
+	}
 
 	return (
 		<div className="">
-			<UserDetails userData = {data}/>
+			<UserDetails userData = {userData}/>
 		</div>
 	)
 }
